[rl] Qwen3.5 GDN cotrain: first end-to-end run + trainer-parity productionization
First time the Qwen3.5 (GDN) port trains end-to-end in the cotrain harness.
Promotes the GDN bitwise-parity work from env-var experiments to a real
generator config, unifies the trainer single-sample path with the packed /
per-request-generator fla+cu path, and turns on validated prefix caching.

- generator: replace the TT_GDN_WRAPPER_EXTERNAL_FLA / TT_GDN_RECURRENT_PREFILL
  env opt-ins with VLLMGenerator.Config.gdn_trainer_parity. When on, the wrapper
  GDN prefill runs the trainer's external fla conv + chunk (bitwise vs trainer)
  and seeds the paged conv_state so decode continues correctly; forces cudagraph
  off (fla decode is not capturable) and prefix cache off (a reused prefix is a
  continuation -> vendored fallback, not bitwise), and raises
  max_num_batched_tokens to max_model_len so fresh prefills are not chunked.
- gdn_vllm_unified: add set_trainer_parity_fla + _seed_conv_state; split forward
  into a trainer-parity path (fresh prefill bitwise; chunked-prefill /
  continuation -> vendored fallback) and the vendored path. Drop the env-var
  branches from gdn_vllm_titan / gdn_vllm_unified.
- model: _cu_seqlens_from_positions returns [0, L] for a single sample so the
  trainer runs the SAME fla+cu path as the per-request generator -- a sample's
  logprobs no longer depend on whether it got packed. TP/SP single-segment falls
  back to non-packed (parity only matters under FSDP, generator TP=1).
- swe_r2e: prefix caching default ON (validated: reused prefix reproduces the
  no-cache output token-for-token + stable async loop; stale state dropped on
  weight sync).
- alphabet_sort unified: mamba_ssm_cache_dtype=float32 to match the trainer's
  fp32 fla recurrent state.
- tmax: add rl_grpo_qwen3_4b_tmax, a DENSE Qwen3-4B batch-invariant control that
  isolates GDN-specific logprob drift from generic batch/kernel nondeterminism.
- dppo: SWE_DEBUG_MAX_LOGDIFF tooling (locate + dump the worst |trainer-gen|
  token with packed-boundary position info); thread positions through the loss
  (DAPO accepts and ignores it for chunked-loss parity).
- training_sample_builder: assert trained-token count == sum(per-turn completion
  lengths) -- TITO / prefix-dedup double-training guard.
- train.py: RL_GPU_OFFSET for local runs on a shared-host GPU subset.
- tests + GDN_BITWISE_PARITY.md updated for the unified single/packed path.

+++ torchtitan/experiments/rl/examples/alphabet_sort/config_registry.py
@@ -235,7 +235,13 @@ def rl_grpo_qwen3_5_4b_varlen_unified() -> Controller.Config:
     cfg = rl_grpo_qwen3_5_4b_varlen()
     return dataclasses.replace(
         cfg,
-        generator=dataclasses.replace(cfg.generator, backend="torchtitan_wrapper"),
+        generator=dataclasses.replace(
+            cfg.generator,
+            backend="torchtitan_wrapper",
+            # VLLMGatedDeltaNetCore requires fp32 ssm state (matches the trainer FLA
+            # recurrent state); vLLM "auto" would follow model dtype (bf16).
+            mamba_ssm_cache_dtype="float32",
+        ),
     )
